+++ packages/mock/src/lib/routes.ts
@@ -1,6 +1,10 @@
 // Aries 1
-
 export const AXI_USER_PERMISSIONS = '/api/internal/v1/axi-user-permissions';
 export const LOG = '/api/v1/log-messages';
 export const REGIONS = '/api/sdk/platform/v1/regions';
 export const SETTINGS = '/api/utils/v1/settings';
+export const ORGANIZATIONS = '/api/sdk/platform/v1/organizations';
+export const USERS = '/api/sdk/platform/v1/users/me';
+
+// Misc
+export const EXAMPLE = '/api/v1/example';