+++ packages/transitions/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.0](https://github.com/Availity/element/compare/@availity/mui-transitions@0.1.2...@availity/mui-transitions@0.2.0) (2024-06-21)
+
+### Dependency Updates
+
+* `mui-typography` updated to version `0.1.2`
+
+### Features
+
+* **mui-transitions:** add Fade component ([0a6e4b9](https://github.com/Availity/element/commit/0a6e4b9b31da786f26f64a66ceef789a57d302fe))
+
 ## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-transitions@0.1.1...@availity/mui-transitions@0.1.2) (2024-04-19)
 
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-transitions@0.1.0...@availity/mui-transitions@0.1.1) (2024-04-04)